backend/requirements.txt
# fastapi==0.95.0
# uvicorn==0.21.0
# sqlalchemy==1.4.46
# python-dotenv==1.0.0
# python-multipart==0.0.6
# python-jose==3.3.0
# passlib==1.7.4
# bcrypt==4.0.1

# ═══════════════════════════════════════════════════════════════════════════
# ARQUIVO: requirements.txt
# DESCRIÇÃO: Dependências do projeto Connect Cidade
# 
# VERSÃO COMPATÍVEL COM PYDANTIC 2.0
# Evita conflito entre FastAPI 0.95.0 e Pydantic 2.0
# ═══════════════════════════════════════════════════════════════════════════

# CORE - Framework Web e Banco de Dados (VERSÕES COMPATÍVEIS)
# ═══════════════════════════════════════════════════════════════════════════
fastapi==0.104.1                   # Framework web (compatível com Pydantic 2.0)
uvicorn[standard]==0.24.0          # Servidor ASGI
sqlalchemy==2.0.23                 # ORM (compatível com Pydantic 2.0)
psycopg2-binary==2.9.11            # Driver PostgreSQL (binary, sem compilação)
python-dotenv==1.0.0               # Carregar variáveis de .env
python-multipart==0.0.6            # Upload de arquivos em FastAPI

# VALIDAÇÃO DE DADOS
# ═══════════════════════════════════════════════════════════════════════════
pydantic[email]==1.10.15             # Validação de dados com email
email-validator==2.1.0             # Validação de email

# SEGURANÇA - JWT, Hash de Senhas
# ═══════════════════════════════════════════════════════════════════════════
python-jose[cryptography]==3.3.0   # JWT tokens
passlib[bcrypt]==1.7.4             # Hash de senhas
cryptography==40.0.0               # Funcionalidades criptográficas
bcrypt==4.0.1                       # Criptografia bcrypt

# UTILITÁRIOS
# ═══════════════════════════════════════════════════════════════════════════
requests==2.31.0                   # HTTP requests
aiofiles==23.2.0                   # Arquivos assíncronos
geopy==2.3.0                       # Geolocalização
Pillow==12.0.0                     # Processamento de imagens (upload, compressão, EXIF)

# NOTAS:
# ═══════════════════════════════════════════════════════════════════════════
# 
# 1. Versões testadas e compatíveis:
#    - FastAPI 0.104.1 ✅ (compatível com Pydantic 2.0)
#    - SQLAlchemy 2.0.23 ✅ (compatível com Pydantic 2.0)
#    - Pydantic 2.0.0 ✅ (com email validation)
#
# 2. Se tiver conflito ao instalar:
#    pip uninstall fastapi pydantic sqlalchemy -y
#    pip install -r requirements.txt
#
# 3. Pillow: REMOVIDO (opcional, só se processar imagens)
#    Se precisar: pip install pillow==10.0.0
#
# ═════════════════════════════════════════════════════════════